Init build details view

+++ AppMovies/AppMovies/Models/MovieDetails/MovieDetails.swift
@@ -29,7 +29,7 @@ struct MovieDetails: Codable {
     let title: String
     let video: Bool
     let voteAverage: Double
-    let voteCount: Double
+    let voteCount: Int
     
     enum CodingKeys: String, CodingKey {
         case adult = "adult"
